+++ lib/core/routers/get_router.dart
@@ -1,14 +1,14 @@
 import 'package:debateseason_frontend_v1/core/routers/get_router_name.dart';
 import 'package:debateseason_frontend_v1/features/auth/bindings/auth_binding.dart';
 import 'package:debateseason_frontend_v1/features/auth/presentation/views/auth_screen.dart';
-import 'package:debateseason_frontend_v1/features/category/bindings/category_binding.dart';
-import 'package:debateseason_frontend_v1/features/category/presentation/views/category_screen.dart';
 import 'package:debateseason_frontend_v1/features/chat/bindings/chat_binding.dart';
 import 'package:debateseason_frontend_v1/features/chat/bindings/debate_binding.dart';
 import 'package:debateseason_frontend_v1/features/chat/presentation/views/chat_room_screen.dart';
 import 'package:debateseason_frontend_v1/features/chat/presentation/views/debate_room_screen.dart';
 import 'package:debateseason_frontend_v1/features/issue/bindings/issue_binding.dart';
 import 'package:debateseason_frontend_v1/features/issue/presentation/views/issue_room_screen.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/bindings/issuemap_binding.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/presentation/views/issuemap_screen.dart';
 import 'package:debateseason_frontend_v1/features/main/main_binding.dart';
 import 'package:debateseason_frontend_v1/features/main/main_screen.dart';
 import 'package:debateseason_frontend_v1/features/profile/bindings/profile_binding.dart';
@@ -65,9 +65,9 @@ class GetRouter {
       binding: IssueBinding(),
     ),
     GetPage(
-      name: GetRouterName.home,
-      page: () => CategoryScreen(),
-      binding: CategoryBinding(),
+      name: GetRouterName.issuemap,
+      page: () => IssuemapScreen(),
+      binding: IssuemapBinding(),
     ),
     GetPage(
       name: GetRouterName.debate,
