+++ lib/core/routers/get_router_name.dart
@@ -3,7 +3,7 @@ class GetRouterName {
   static const String splash = '/';
   static const String auth = '/login';
   static const String main = '/main';
-  static const String home = '/home';
+  static const String issuemap = '/issuemap';
   static const String chat = '/chat';
   static const String issue = '/issue';
   static const String category = '/category';
