+++ lib/features/issuemap/category_constants.dart
