+++ lib/features/issuemap/data/mapper/category_mapper.dart
@@ -1,5 +1,5 @@
-import 'package:debateseason_frontend_v1/features/category/data/models/remote/response/users_home_res.dart';
-import 'package:debateseason_frontend_v1/features/category/domain/entities/category_entity.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/data/models/remote/response/users_home_res.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/domain/entities/category_entity.dart';
 
 class CategoryMapper {
   CategoryEntity toEntity(UsersHomeRes res) {
