+++ lib/features/issuemap/domain/entities/category_entity.freezed.dart
