+++ lib/features/issuemap/domain/repositories/remote/users_home_repository.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/category/domain/entities/category_entity.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/domain/entities/category_entity.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 
 abstract class UsersHomeRepository {
