+++ lib/features/issuemap/presentation/view_models/issuemap_view_model.dart
@@ -1,9 +1,9 @@
-import 'package:debateseason_frontend_v1/features/category/domain/entities/category_entity.dart';
-import 'package:debateseason_frontend_v1/features/category/domain/repositories/remote/users_home_repository.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/domain/entities/category_entity.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/domain/repositories/remote/users_home_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 import 'package:get/get.dart';
 
-class CategoryViewModel extends GetxController {
+class IssuemapViewModel extends GetxController {
   late final UsersHomeRepository _usersHomeRepository;
   final _categories =
       Rx<UiState<List<CategoryEntity>>>(const UiState.loading());
