+++ lib/features/main/main_binding.dart
@@ -1,5 +1,5 @@
-import 'package:debateseason_frontend_v1/features/category/bindings/category_binding.dart';
 import 'package:debateseason_frontend_v1/features/home/home_binding.dart';
+import 'package:debateseason_frontend_v1/features/issuemap/bindings/issuemap_binding.dart';
 import 'package:debateseason_frontend_v1/features/profile/bindings/profile_binding.dart';
 import 'package:get/get.dart';
 
@@ -10,7 +10,7 @@ class MainBinding extends Bindings {
   void dependencies() {
     Get.lazyPut(() => MainViewModel());
     HomeBinding().dependencies();
-    CategoryBinding().dependencies();
+    IssuemapBinding().dependencies();
     ProfileBinding().dependencies();
   }
 }
