+++ lib/features/profile/presentation/views/profile_input_page.dart
@@ -87,7 +87,7 @@ class ProfileInputPage extends GetView<ProfileInputViewModel> {
                                 controller.setApiLoading(isApiLoading: true);
                               }, success: (message) {
                                 controller.setApiLoading(isApiLoading: false);
-                                Get.offAllNamed(GetRouterName.home);
+                                Get.offAllNamed(GetRouterName.issuemap);
                               }, failure: (msg) {
                                 controller.setApiLoading(isApiLoading: false);
                                 deSnackBar(msg);
